refactor upcoming lessons
