+++ SchoolManagement.Application/Features/Lesson/Dtos/LessonDto.cs
@@ -6,6 +6,7 @@ public class LessonDto
 {
     public Guid Id { get; set; } 
     public string TeacherId { get; set; } = string.Empty;
+    public Guid classRoomId { get; set; } 
     public string Title { get; set; } = string.Empty;
     public string Grade { get; set; } = string.Empty;
     public string Subject { get; set; } = string.Empty;
